Date filtering for mts objects

Subsets a mts object by date. This function always filters to day-boundaries. For sub-day filtering, use mts_filterDatetime().

Dates can be anything that is understood by MazamaCoreUtils::parseDatetime() including either of the following recommended formats:

"YYYYmmdd"

"YYYY-mm-dd"

Timezone determination precedence assumes that if you are passing in POSIXct values then you know what you are doing.

get timezone from startdate if it is POSIXct

use passed in timezone

get timezone from mts

mts_filterDate(
  mts = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  unit = "sec",
  ceilingStart = FALSE,
  ceilingEnd = FALSE
)

Arguments

mts
mts object.

startdate
Desired start datetime (ISO 8601).

enddate
Desired end datetime (ISO 8601).

timezone
Olson timezone used to interpret dates.

unit
Units used to determine time at end-of-day.

ceilingStart
Logical instruction to apply ceiling_date to the startdate rather than floor_date

ceilingEnd
Logical instruction to apply ceiling_date to the enddate rather than floor_date

Value

A subset of the given mts object.

Note

The returned data will run from the beginning of startdate until the beginning of enddate -- i.e. no values associated with enddate will be returned. The exception being when enddate is less than 24 hours after startdate. In that case, a single day is returned.

See also

Examples

library(MazamaTimeSeries)

example_mts %>%
  mts_filterDate(startdate = 20190703, enddate = 20190706) %>%
  mts_extractData() %>%
  head()

#>              datetime 4af81ed0f6ec7acd_4734 c3ac53faf335f19f_5030
#> 1 2019-07-03 07:00:00                    NA                    NA
#> 2 2019-07-03 08:00:00                    NA                    NA
#> 3 2019-07-03 09:00:00                    NA                    NA
#> 4 2019-07-03 10:00:00                    NA                    NA
#> 5 2019-07-03 11:00:00                    NA                    NA
#> 6 2019-07-03 12:00:00                    NA                    NA
#>   776e00b070251a36_23237 ec61b544c1dbfe21_2448 70e599f772b22877_4670
#> 1                     NA                    NA              17.08546
#> 2                     NA                    NA              12.64041
#> 3                     NA                    NA              10.28893
#> 4                     NA                    NA              13.02315
#> 5                     NA                    NA              18.03966
#> 6                     NA                    NA              16.85280
#>   e0d610f7f219e39f_5080 2ee691d37eaa1aff_3529 ecd1d2c35c1c18cf_3505
#> 1                    NA                    NA                    NA
#> 2                    NA                 11.55                    NA
#> 3              4.753222                    NA                    NA
#> 4                    NA                    NA                    NA
#> 5                    NA                    NA                    NA
#> 6                    NA                    NA                    NA
#>   05533d5b998ac7bf_3525 91299b887fface5f_3515 c316bcf7a85c8242_3617
#> 1                    NA                    NA              16.51333
#> 2               11.7100                    NA              13.03900
#> 3                    NA                    NA              10.66850
#> 4               12.7475              16.21471              13.89267
#> 5               16.4975              20.20000              18.34569
#> 6               16.3500                    NA              18.56433
#>   a812772f2ef60b6f_3527 70f70265dbeb7f36_3551 f1eb28176ffa3c00_3487
#> 1                    NA              17.39217              15.65583
#> 2                    NA              13.85500              12.32800
#> 3                    NA              11.89383              10.97917
#> 4                    NA              17.52333              16.08800
#> 5                    NA              21.10167              19.23083
#> 6                    NA              19.86083              16.80583
#>   28a43b2d6854838f_4725 da4cadd2d6ea5302_4686 d68d5d33331df58f_4689
#> 1              15.54783             13.701500              15.48167
#> 2              12.13850             10.981500              12.08617
#> 3              10.13750              9.594667              10.23333
#> 4              11.01800             14.201333              13.97850
#> 5              16.42133             15.973500              17.96500
#> 6              17.25483             14.861333              16.37350
#>   2de587fc7359dd00_4662 07702ba2209dec22_4702 1fb48f53c052893a_4692
#> 1                    NA              18.14983             13.901500
#> 2                    NA              12.70850             10.916000
#> 3                    NA              10.05633              9.147667
#> 4                    NA              11.67983              9.871833
#> 5                    NA              16.64967             14.488000
#> 6                    NA              19.47883             14.453000
#>   ada194ce98ce4bed_5184 a3eb0d42e5c78f8b_5228 065b6608aa9e6136_5218
#> 1              15.80300              13.96083              16.50865
#> 2              12.40483              11.12050              13.03933
#> 3              10.15850               9.28400              10.84050
#> 4              12.21683              13.89650              11.86569
#> 5              15.76917              15.72600              17.53732
#> 6              17.91300              13.85217              20.55909
#>   36fa039140645de8_2504 173ff64a55da1183_2693 055497925c615bbd_2452
#> 1              28.38150              21.78717                    NA
#> 2              15.79533              14.02233                    NA
#> 3              14.22517              12.42267                    NA
#> 4              16.94700              13.94450                    NA
#> 5              18.13700              14.96133                    NA
#> 6              18.30750              15.49350                    NA
#>   6db0b260ed58bea0_2713 8d9ad84c05e66fcb_2496 30b3a848f934811d_4746
#> 1              26.07433              21.42550                    NA
#> 2              15.22517              14.11983                    NA
#> 3              13.10017              11.44983                    NA
#> 4              15.65950              14.92733                    NA
#> 5              16.66467              18.45717                    NA
#> 6              18.81417              21.68833                    NA
#>   bf7350cd6c2ec4c1_4727 452fd84452485d47_23233 bde3b79da492ff98_23429
#> 1                    NA               14.93218              15.670000
#> 2                    NA               12.01389              12.404833
#> 3                    NA               10.17417               9.857667
#> 4                    NA               13.17920              10.399833
#> 5                    NA               18.20469              16.129167
#> 6                    NA               17.11636              18.473667
#>   76a49e0074377eb5_23427 8c345724ee28a6ef_23253 85a2219173d7e0e6_22695
#> 1               19.21583               14.94367                     NA
#> 2               15.76250               11.48950                     NA
#> 3               13.02417               10.04633                     NA
#> 4               14.52617               14.05750                     NA
#> 5               20.66267               18.30517                     NA
#> 6               26.17500               16.02133                     NA
#>   f4b229bfc8882d90_22687 dfed9bdf5bbdeee0_23093 d76f0e830185fe4c_23291
#> 1                     NA                     NA                     NA
#> 2                     NA                     NA                     NA
#> 3                     NA                     NA                     NA
#> 4                     NA                     NA                     NA
#> 5                     NA                     NA                     NA
#> 6                     NA                     NA                     NA