Calculate distances from mts locations to a location of interest

This function uses the geodist package to return the distances (meters) between mts locations and a location of interest. These distances can be used to create a mask identifying monitors within a certain radius of the location of interest.

mts_getDistance(
  mts = NULL,
  longitude = NULL,
  latitude = NULL,
  measure = c("geodesic", "haversine", "vincenty", "cheap")
)

Arguments

mts
mts object.

longitude
Longitude of the location of interest.

latitude
Latitude of the location of interest.

measure
One of "geodesic", "haversine", "vincenty" or "cheap"

Value

Vector of of distances (meters) named by deviceDeploymentID.

Note

The measure "cheap" may be used to speed things up depending on the spatial scale being considered. Distances calculated with measure = "cheap" will vary by a few meters compared with those calculated using measure = "geodesic".

Examples

library(MazamaTimeSeries)

# Garfield Medical Center in LA
longitude <- -118.12321
latitude <- 34.06775

distances <- mts_getDistance(
  mts = example_mts,
  longitude = longitude,
  latitude = latitude
)

# Which sensors are within 1000 meters of Garfield Med Ctr?
distances[distances <= 1000]

#>  ec61b544c1dbfe21_2448  2ee691d37eaa1aff_3529  ecd1d2c35c1c18cf_3505 
#>               761.2503               844.0852               743.9420 
#>  05533d5b998ac7bf_3525 8c345724ee28a6ef_23253 
#>               765.9372               950.1779