Required columns for the 'meta' dataframe

The 'meta' dataframe found in sts and mts objects is required to have a minimum set of information for proper functioning of the package. The names of these columns are specified in requiredMetaNames and include:

deviceDeploymentID -- unique identifier (see MazmaLocationUtils)

deviceID -- device identifier

locationID -- location identifier (see MazmaLocationUtils)

locationName -- English language name

longitude -- decimal degrees E

latitude -- decimal degrees N

elevation -- elevation of station in m

countryCode -- ISO 3166-1 alpha-2

stateCode -- ISO 3166-2 alpha-2

timezone -- Olson time zone

requiredMetaNames

Format

A vector with 10 elements

Details

requiredMetaNames